requirements.txt
requests>=2.31.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0